LIM_long_mode/requirements.txt
numpy>=1.19.5
pandas>=1.3.5
matplotlib>=3.5.1
seaborn>=0.11.2
scikit-learn>=1.0.2
tensorflow>=2.8.0
tensorflow-probability>=0.16.0
gym>=0.21.0
talib-binary>=0.4.19
tushare>=1.2.62
PyYAML>=6.0 